VCV-150: Refactor Annotation feature
