+++ src/features/annotation/task-details/components/workspace/specimen-image-viewer.tsx
