+++ src/features/annotation/tasks-list/components/tasks/annotation-tasks-list.tsx
@@ -5,8 +5,8 @@ import type { GetAnnotationTasksQueryParams } from '@/api/annotation-task/valida
 import { type AnnotationTaskStatus } from '@/api/annotation-task/validation/annotation-task-schema';
 import { Fragment, useEffect } from 'react';
 import { usePagination } from '@/lib/hooks/use-pagination';
-import AnnotationTaskCard from '@/features/annotation/components/tasks-list/annotation-task-card';
-import AnnotationTasksPagination from '@/features/annotation/components/tasks-list/annotation-tasks-pagination';
+import AnnotationTaskCard from '@/features/annotation/tasks-list/components/tasks/annotation-task-card';
+import AnnotationTasksPagination from '@/features/annotation/tasks-list/components/layout/annotation-tasks-pagination';
 import { Separator } from '@/components/ui/separator';
 
 interface AnnotationTasksListProps {
